requirements.txt
# Data processing
pandas==2.1.0
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# Machine Learning
scikit-learn==1.3.0
lightgbm==4.0.0
catboost==1.2.2
xgboost==1.7.6
bayesian-optimization==1.4.3
shap==0.42.1

# Deep Learning & GNNs
torch==2.0.1
torch-geometric==2.3.1
dgl==1.1.2

# Bayesian Modeling
pymc==5.7.2
arviz==0.16.1

# API & Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
alembic==1.12.0

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Data Sources & APIs
requests==2.31.0
aiohttp==3.8.5
yfinance==0.2.20
alpha-vantage==2.3.1
quandl==3.7.0

# ETL & Orchestration
apache-airflow==2.7.0
prefect==2.11.5
dbt-core==1.6.0
dbt-postgres==1.6.0

# Cloud & Storage
boto3==1.28.25
azure-storage-blob==12.17.0
google-cloud-storage==2.10.0
s3fs==2023.8.0

# LLM & AI
openai==0.28.0
langchain==0.0.291
chromadb==0.4.8
sentence-transformers==2.2.2
tiktoken==0.4.0

# Visualization
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.6
tqdm==4.65.0
slugify==0.0.1
